alarm_system/requirements.txt
flask
mtcnn
pillow
numpy
tensorflow
torch